Replacing to loop and adding focus
1. Although the previous solution worked, surrounding with while for
readability.
2. Focus ( if that's the name ) :
 - First, reading the exsiting size value from the .toml
 - Then, on load, the focus will be on the configured font size, rather
than on the first item. Changing font size will keep the focus on the
new value.

* Inside the while, in the first run, the sed will rewrite the existing
value; the other approach is to move this line below the choise
selection, but then, there's a need to first check that in case of
"<<Back" the value won't accidently change the size = << Back in the
.toml.
To avoid it, I just moved the sed line above the choice and know that in
the first run it will override the same value in the .toml; still looks
cleaner than adding an if or a case loop.