Defaulting file permission to umask 022
Default is 777.
Currently, after omakub installation, umask is 002(run umask in
thebash); which defaults directories to 775 (777-2) and files to
664(666-2).
With umask 022; directories will be 755 (777-22) and files 644 (666-22).
This will set directories permission to:owner to rwx, while group
andother r_x.And file permissions to: owner to rw_, while group and
other r__.

Git is making such changes automatically, behind the scenes
whenuploading a file or directory, and I set my directories and files in
thesame manner so I think it's correct.