feat: install gum manually with out using apt
The main benifit of using apt is that you can upgrade etc, but because
we are manually fetching the package and pushing it into the install
feature, you cant have apt do the upgrades later anyway.

I guess one reason to keep it the way that it is, is if it handles
variations in cpu architectures as this package is an x86 binary.

The reason for this change is i am making a fork for fedora and the
would like to keep as much of the code the same were possible